AERN2.Poly.Basics

Description

Basics of unary sparse polynomials

Documentation

type PolyCoeffRing c = (Ring c, HasIntegers c, HasAccuracy c, HasNorm c, Show c)

A shortcut type constraint for types suitable as coefficients of our polynomials, loose enough to permit Integer coefficients.

type PolyCoeffField c = (PolyCoeffRing c, Field c, HasDyadics c, CanAddSubMulDivBy c Dyadic)

A shortcut type constraint for types suitable as coefficients of our polynomials, loose enough to permit Rational coefficients.

type PolyCoeffBall c = (PolyCoeffField c, CanAddSubMulDivBy c CauchyReal, IsInterval c c, IsBall c, CanSetPrecision c)

a shortcut type constraint for types suitable as coefficients of our polynomials

data Poly c

Constructors

Poly

Fields

poly_terms :: Terms c

type Degree = Integer

type Terms c = Map Degree c

terms_empty :: Terms c

terms_size :: Terms c -> Integer

terms_insertWith :: (c -> c -> c) -> Degree -> c -> Terms c -> Terms c

terms_toList :: Terms c -> [(Degree, c)]

terms_toDescList :: Terms c -> [(Degree, c)]

terms_fromList :: HasIntegers c => [(Degree, c)] -> Terms c

terms_fromListAddCoeffs :: (CanAddSameType c, HasIntegers c) => [(Degree, c)] -> Terms c

terms_unionWith :: (c -> c -> c) -> Terms c -> Terms c -> Terms c

terms_map :: (c1 -> c2) -> Terms c1 -> Terms c2

terms_filterKeepConst :: (Degree -> c -> Bool) -> Terms c -> Terms c

terms_filterMayLoseConst :: (Degree -> c -> Bool) -> Terms c -> Terms c

terms_degree :: Terms c -> Degree

terms_degrees :: Terms c -> [Degree]

terms_coeffs :: Terms c -> [c]

terms_updateConst :: HasIntegers c => (c -> c) -> Terms c -> Terms c

terms_lookupCoeff :: HasIntegers c => Terms c -> Degree -> c

terms_lookupCoeffDoubleConstTerm :: (HasIntegers c, CanAddSameType c) => Terms c -> Degree -> c

formatTerms :: HasIntegers c => (c -> (String, Bool, Bool)) -> Terms c -> String