AERN2.Poly.Cheb.DCT

Description

Interpolation using Discrete cosine transform

Documentation

lift1_DCT

Arguments

:: (Field c, CanMulBy c CauchyReal, CanNormalize (ChPoly c), Show c)

=> (Degree -> Degree)
detemining a degree bound for the result from the degree of p

-> (c -> c)
the function f to apply pointwise to p

-> ChPoly c
p

-> ChPoly c

DCT-approximate the result of applying the given function f pointwise to the given polynomial p.

lift2_DCT

Arguments

:: (PolyCoeffBall c, CanNormalize (ChPoly c))

=> (Degree -> Degree -> Degree)
detemining a degree bound for the result from the degrees of p1 and p2

-> (c -> c -> c)
the function f to apply pointwise to p1 and p2

-> ChPoly c
p1

-> ChPoly c
p2

-> ChPoly c

DCT-approximate the result of applying the given binary function f pointwise to the given polynomials p1 and p2.