AERN2.RealFun.Operations

Description

Classes for real number function operations

Documentation

class HasDomain f where

Minimal complete definition

getDomain

Associated Types

type Domain f

Methods

getDomain :: f -> Domain f

data SameDomFnPair f

Constructors

SameDomFnPair (f, f)

class HasDomain f => ArbitraryWithDom f where

Minimal complete definition

arbitraryWithDom

Methods

arbitraryWithDom :: Domain f -> Gen f

class CanApply f x where

Minimal complete definition

apply

Associated Types

type ApplyType f x

Methods

apply :: f -> x -> ApplyType f x

compute f(x)

class CanApplyApprox f x where

Give an unsafe etimate of the function's range which is fast to compute. Intended to be used in optimisation heuristics.

Minimal complete definition

applyApprox

Associated Types

type ApplyApproxType f x

Methods

applyApprox :: f -> x -> ApplyApproxType f x

compute a cheap and unsafe approximation of f(x)

sampledRange :: (CanApply f t, ApplyType f t ~ t, CanMinMaxSameType t, ConvertibleExactly Dyadic t, Show t) => DyadicInterval -> Integer -> f -> Interval t t

Evaluate a function on a regular grid of the given size and return the largerst and smallest values found. Useful for making instances of class CanApplyApprox.

class HasVars f where

Minimal complete definition

varFn

Associated Types

type Var f

Methods

varFn :: f -> Var f -> f

the function x, ie the function that project the domain to the given variable x

specEvalUnaryVarFn :: (HasVars f, Var f ~ (), HasDomain f, CanMapInside (Domain f) x, CanApply f x, HasEqCertainly x (ApplyType f x), Arbitrary f, Arbitrary x, Show f, Show x) => T f -> T x -> Spec

type HasConstFunctions t f = (HasDomain f, ConvertibleExactly (Domain f, t) f)

constFn :: HasConstFunctions t f => (Domain f, t) -> f

specEvalConstFn :: (HasConstFunctions c f, CanMapInside (Domain f) x, CanApply f x, HasEqCertainly c (ApplyType f x), Arbitrary c, Arbitrary f, Arbitrary x, Show c, Show f, Show x) => T c -> T f -> T x -> Spec

class CanMaximiseOverDom f d where

Minimal complete definition

maximumOverDom

Associated Types

type MaximumOverDomType f d

Methods

maximumOverDom :: f -> d -> MaximumOverDomType f d

class CanMinimiseOverDom f d where

Minimal complete definition

minimumOverDom

Associated Types

type MinimumOverDomType f d

Methods

minimumOverDom :: f -> d -> MinimumOverDomType f d

specCanMaximiseOverDom :: (HasDomain f, CanMapInside (Domain f) x, CanApply f x, ApplyType f x ~ v, CanMaximiseOverDom f (Domain f), HasOrderCertainly v (MaximumOverDomType f (Domain f)), Arbitrary f, Show f, Arbitrary x, Show x) => T f -> T x -> Spec

class CanIntegrateOverDom f bounds where

Minimal complete definition

integrateOverDom

Associated Types

type IntegralOverDomType f bounds

Methods

integrateOverDom :: f -> bounds -> IntegralOverDomType f bounds