AERN2.RealFun.SineCosine

Description

Pointwise sine and cosine for functions

Documentation

sineWithAccuracyGuide :: (HasDomain f, CanMaximiseOverDom f (Domain f), CanMinimiseOverDom f (Domain f), MinimumOverDomType f (Domain f) ~ MaximumOverDomType f (Domain f), IsInterval MPBall (MinimumOverDomType f (Domain f)), CanNegSameType f, CanAddSameType f, CanMulSameType f, CanAddSubMulDivCNBy f Integer, CanAddSubMulDivCNBy f CauchyReal, HasAccuracy f, CanSetAccuracyGuide f, CanSetPrecision f, CanReduceSizeUsingAccuracyGuide f, IsBall f, Show f) => Accuracy -> f -> f

cosineWithAccuracyGuide :: (HasDomain f, CanMaximiseOverDom f (Domain f), CanMinimiseOverDom f (Domain f), MinimumOverDomType f (Domain f) ~ MaximumOverDomType f (Domain f), IsInterval MPBall (MinimumOverDomType f (Domain f)), CanNegSameType f, CanAddSameType f, CanMulSameType f, CanAddSubMulDivCNBy f Integer, CanAddSubMulDivCNBy f CauchyReal, HasAccuracy f, CanSetAccuracyGuide f, CanSetPrecision f, CanReduceSizeUsingAccuracyGuide f, IsBall f, Show f) => Accuracy -> f -> f

sineCosineWithAccuracyGuide :: (HasDomain f, CanMaximiseOverDom f (Domain f), CanMinimiseOverDom f (Domain f), MinimumOverDomType f (Domain f) ~ MaximumOverDomType f (Domain f), IsInterval MPBall (MinimumOverDomType f (Domain f)), CanNegSameType f, CanAddSameType f, CanMulSameType f, CanAddSubMulDivCNBy f Integer, CanAddSubMulDivCNBy f CauchyReal, HasAccuracy f, CanSetAccuracyGuide f, CanSetPrecision f, CanReduceSizeUsingAccuracyGuide f, IsBall f, Show f) => Bool -> Accuracy -> f -> f

sineTaylorSum :: (CanAddSameType f, CanMulSameType f, CanAddSubMulDivCNBy f Integer, HasAccuracy f, CanSetPrecision f, CanReduceSizeUsingAccuracyGuide f, Show f) => (Accuracy -> f) -> MPBall -> Accuracy -> (f, ErrorBound, Integer)

For a given polynomial p, compute a partial Taylor sum of cos(p) and return it together with its error bound e and the degree of the polynomial n.

cosineTaylorSum :: (CanAddSameType f, CanMulSameType f, CanAddSubMulDivCNBy f Integer, HasAccuracy f, CanSetPrecision f, CanReduceSizeUsingAccuracyGuide f, Show f) => (Accuracy -> f) -> MPBall -> Accuracy -> (f, ErrorBound, Integer)

For a given polynomial p, compute a partial Taylor sum of cos(p) and return it together with its error bound e and the degree of the polynomial n.

sineCosineTaylorSum :: (CanAddSameType f, CanMulSameType f, CanAddSubMulDivCNBy f Integer, HasAccuracy f, CanSetPrecision f, CanReduceSizeUsingAccuracyGuide f, Show f) => Bool -> (Accuracy -> f) -> MPBall -> Accuracy -> (f, ErrorBound, Integer)