AERN2.RealFun.UnaryModFun

Description

Real functions by evaluator on dyadic points and a modulus of continuity

Documentation

data UnaryModFun

Here we represent a continuous function \(f\) using its evaluation function on dyadics and a "local" modulus of continuity \(\omega\) in the following sense: . For any dyadic \(d_1,d_2\in b\) with \(|d_1-d_2|\leq 2^{ -\omega(b)(i)}\) we have \(|f(d_1)-f(d_2)| \leq 2^{ -i }\).

For any \(b\), the modulus of continuity has to be a non-decreasing function: \[ i\leq j \implies \omega(b)(i)\leq\omega(b)(j) \] and \(omega(b)(i)\) has to converge to \(\infty\) with increasing \(i\).

Constructors

UnaryModFun

Fields

_modfun_domain :: DyadicInterval

_modfun_eval :: CatchingNumExceptions Dyadic -> CatchingNumExceptions CauchyReal

\(d mapsto f(d)\)

_modfun_modulus :: MPBall -> Integer -> Integer

\(omega\)

unaryModFun :: ConvertibleExactly t UnaryModFun => t -> UnaryModFun

inverseNonDecreasingFnAnyBelow

Arguments

:: (Integer -> Integer)
\(f\)

-> Integer -> Integer

For a monotone integer function \(f\), and an integer \(n\) which is neither below or above the range of \(f\), return the largest \(i\) such that \(f(i) \leq n\).

inverseNonDecreasingFnMaxBelow

Arguments

:: (Integer -> Integer)
\(f\)

-> Integer -> Integer

For a monotone integer function \(f\), and an integer \(n\) which is neither below or above the range of \(f\), return the largest \(i\) such that \(f(i) \leq n\).