AERN2.Poly.Basics

Description

Basics of unary sparse polynomials

Documentation

class (Ring c, HasIntegers c, HasAccuracy c, HasNorm c, Show c) => PolyCoeffRing c

An aggregate sub-class for types suitable as coefficients of our polynomials, loose enough to permit Integer coefficients.

class (PolyCoeffRing c, Field c, HasDyadics c, CanAddSubMulDivCNBy c Dyadic) => PolyCoeffField c

An aggregate sub-class for types suitable as coefficients of our polynomials, loose enough to permit Rational coefficients.

class (PolyCoeffField c, CanAddSubMulDivCNBy c CauchyReal, IsInterval c c, IsBall c, CanSetPrecision c) => PolyCoeffBall c

An aggregate sub-class for types suitable as coefficients of our polynomials

newtype Poly c

Constructors

Poly

Fields

poly_terms :: Terms c

type Degree = Integer

type Terms c = Map Degree c

terms_empty :: Terms c

terms_size :: Terms c -> Integer

terms_insertWith :: (c -> c -> c) -> Degree -> c -> Terms c -> Terms c

terms_toList :: Terms c -> [(Degree, c)]

terms_toDescList :: Terms c -> [(Degree, c)]

terms_fromList :: HasIntegers c => [(Degree, c)] -> Terms c

terms_fromListAddCoeffs :: (CanAddSameType c, HasIntegers c) => [(Degree, c)] -> Terms c

terms_unionWith :: (c -> c -> c) -> Terms c -> Terms c -> Terms c

terms_map :: (c1 -> c2) -> Terms c1 -> Terms c2

terms_filterKeepConst :: (Degree -> c -> Bool) -> Terms c -> Terms c

terms_filterMayLoseConst :: (Degree -> c -> Bool) -> Terms c -> Terms c

terms_degree :: Terms c -> Degree

terms_degrees :: Terms c -> [Degree]

terms_coeffs :: Terms c -> [c]

terms_updateConst :: HasIntegers c => (c -> c) -> Terms c -> Terms c

terms_lookupCoeff :: HasIntegers c => Terms c -> Degree -> c

terms_lookupCoeffDoubleConstTerm :: (HasIntegers c, CanAddSameType c) => Terms c -> Degree -> c

formatTerms :: HasIntegers c => (c -> (String, Bool, Bool)) -> Terms c -> String

Orphan instances

SuitableForCE es => CanExtractCE es (Map Degree)

Methods

extractCE :: CanEnsureCE es c => Maybe es -> Map Degree c -> CollectErrors es (Map Degree (EnsureNoCE es c))