AERN2.Poly.Cheb.Type

Description

Chebyshev basis unary sparse polynomials

Documentation

type ChPolyMB = ChPoly MPBall

data ChPoly c

Constructors

ChPoly

Fields

chPoly_dom :: DyadicInterval

chPoly_poly :: Poly c

chPoly_acGuide :: Accuracy

chPoly_maybeBounds :: Maybe (ChPolyBounds c)

chPoly_terms :: ChPoly c -> Terms c

data ChPolyBounds c

Constructors

ChPolyBounds

Fields

chPolyBounds_lip :: c

chPoly_maybeLip :: ChPoly c -> Maybe c

chPoly_setLip :: c -> ChPoly c -> ChPoly c

type CanBeChPoly c t = ConvertibleExactly t (ChPoly c)

chPoly :: CanBeChPoly c t => t -> ChPoly c

chPolyMPBall :: ConvertibleExactly t (ChPoly MPBall) => t -> ChPoly MPBall

showInternals :: Show c => ChPoly c -> String

fromDomToUnitInterval :: CanAddSubMulDivCNBy t Dyadic => DyadicInterval -> t -> t

serialise :: ChPolyMB -> String

deserialise :: ByteString -> Maybe ChPolyMB

sweepUsingOwnAccuracyGuide :: _ => ChPoly c -> ChPoly c

type Degree = Integer

degree :: ChPoly c -> Integer

reduceDegree :: _ => Degree -> ChPoly c -> ChPoly c

Drop all terms that whose degree is above the given limit or whose norm is at or below the threshold. Compensate for the drops in the constant term.

reduceDegreeWithLostAccuracyLimit :: _ => Accuracy -> ChPoly c -> ChPoly c